==> 1_Water_Inputs.ipynb <==
# Climate and water data from URBAG MAP

This notebook quantifies the green, blue and water extraction per polygon using climate data from the URBAG map.

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns

#In this cell the scenario is chosen. This can either be defined here or in the RUN_ALL.ipynb
#If defined here unmark the scen = line and select the scenario you will run.
#If defined in RUN_ALL.ipynb unmark the %store -r scen line.
#Scenario names are:
#S0_MinFert > N,P,K demand met by mineral fertilizers only
#S0_struvite_P > P demand met by struvite, N demand met by struvite and mineral fertilizer and K demand met by mineral fertilizer
#S0_compost > N,P,K supplied by compost produced in the AMB, remaining N,P,K demand met by mineral fertilizer
#S0_Ammon_salts > N supplied from recovered ammonium salts and from mineral fertilizers, P from struvite and K from mineral fertilizer
scen = 'S0_MinFert'
#%store -r scen

print(scen)

#This cell determines the path and name of the map file
filename = "S4_NewVoronoi_0524.csv"
file_path = f"inputs/{filename}"

URBAG_map = pd.read_csv(file_path)

#Number of plots. Currently 4581 (NPK_27_06_22An - 3_02_23.xlsx)
len_URBAG_map= len(URBAG_map['Voronoi_1'])
len_URBAG_map

URBAG_map.columns

URBAG_map.keys()

#Checking precipitation and evapotranspiration data per month
Precipitation_months = ['01_PPT', '02_PPT', '03_PPT', '04_PPT', '05_PPT', '06_PPT','07_PPT', '08_PPT', '09_PPT', '10_PPT', '11_PPT', '12_PPT']
Evapotranspiration_months = ['01_ETO', '02_ETO', '03_ETO', '04_ETO', '05_ETO', '06_ETO','07_ETO', '08_ETO', '09_ETO', '10_ETO', '11_ETO', '12_ETO']
months = ["Jan","Feb","Mar","Apr","May","Jun","Jul","Aug","Sep","Oct","Nov","Dec",]

type(URBAG_map['03_PPT'][1])

Precipitation_ranges = URBAG_map.agg({'01_PPT' : ['min', 'max'],
                                        '02_PPT' : ['min', 'max'],
                                        '03_PPT' : ['min', 'max'],
                                        '04_PPT' : ['min', 'max'],
                                        '05_PPT' : ['min', 'max'],
                                        '06_PPT' : ['min', 'max'],
                                        '07_PPT' : ['min', 'max'],
                                        '08_PPT' : ['min', 'max'],
                                        '09_PPT' : ['min', 'max'],
                                        '10_PPT' : ['min', 'max'],
                                        '11_PPT' : ['min', 'max'],
                                        '12_PPT' : ['min', 'max']})

Evapotranspiration_ranges = URBAG_map.agg({'01_ETO' : ['min', 'max'],
                                        '02_ETO' : ['min', 'max'],
                                        '03_ETO' : ['min', 'max'],
                                        '04_ETO' : ['min', 'max'],
                                        '05_ETO' : ['min', 'max'],
                                        '06_ETO' : ['min', 'max'],
                                        '07_ETO' : ['min', 'max'],
                                        '08_ETO' : ['min', 'max'],
                                        '09_ETO' : ['min', 'max'],
                                        '10_ETO' : ['min', 'max'],
                                        '11_ETO' : ['min', 'max'],
                                        '12_ETO' : ['min', 'max']})

Kc_ranges = URBAG_map.agg({'Kc_b_Kc1' : ['min', 'max'],
                            'Kc_b_Kc2' : ['min', 'max'],
                            'Kc_b_Kc3' : ['min', 'max'],
                            'Kc_b_Kc4' : ['min', 'max'],
                            'Kc_b_Kc5' : ['min', 'max'],
                            'Kc_b_Kc6' : ['min', 'max'],
                            'Kc_b_Kc7' : ['min', 'max'],
                            'Kc_b_Kc8' : ['min', 'max'],
                            'Kc_b_Kc9' : ['min', 'max'],
                            'Kc_b_Kc10' : ['min', 'max'],
                            'Kc_b_Kc11' : ['min', 'max'],
                            'Kc_b_Kc12' : ['min', 'max']})

Eficiencies_ranges = URBAG_map.agg({'EFIC_DISTR' : ['min', 'max'],
                                    'EFIC_RIEGO' : ['min', 'max']})

Precipitation_ranges

Evapotranspiration_ranges

Kc_ranges

Eficiencies_ranges

## CWR

Crop water requirements (CWR) according to Allen et al. (1998) (mm) per month (m) and farm/plot (f)

    CWRf,m= ETof,m * Kcc,m

URBAG_map["01_CWR"] = URBAG_map["01_ETO"]*URBAG_map["Kc_b_Kc1"]
URBAG_map["02_CWR"] = URBAG_map["02_ETO"]*URBAG_map["Kc_b_Kc2"]
URBAG_map["03_CWR"] = URBAG_map["03_ETO"]*URBAG_map["Kc_b_Kc3"]
URBAG_map["04_CWR"] = URBAG_map["04_ETO"]*URBAG_map["Kc_b_Kc4"]
URBAG_map["05_CWR"] = URBAG_map["05_ETO"]*URBAG_map["Kc_b_Kc5"]
URBAG_map["06_CWR"] = URBAG_map["06_ETO"]*URBAG_map["Kc_b_Kc6"]
URBAG_map["07_CWR"] = URBAG_map["07_ETO"]*URBAG_map["Kc_b_Kc7"]
URBAG_map["08_CWR"] = URBAG_map["08_ETO"]*URBAG_map["Kc_b_Kc8"]
URBAG_map["09_CWR"] = URBAG_map["09_ETO"]*URBAG_map["Kc_b_Kc9"]
URBAG_map["10_CWR"] = URBAG_map["10_ETO"]*URBAG_map["Kc_b_Kc10"]
URBAG_map["11_CWR"] = URBAG_map["11_ETO"]*URBAG_map["Kc_b_Kc11"]
URBAG_map["12_CWR"] = URBAG_map["12_ETO"]*URBAG_map["Kc_b_Kc12"]

np.sort(pd.unique(URBAG_map["01_CWR"]))

## PPTe (Effective precipitation)

Effective precipitation (PPTe) (mm) per month per plot = IF PPT > 4mm then, PPT*0.75, else 0

URBAG_map['01_PPTe'] = np.where(URBAG_map['01_PPT'] >= 4, URBAG_map['01_PPT']*0.75, 0)
URBAG_map['02_PPTe'] = np.where(URBAG_map['02_PPT'] >= 4, URBAG_map['02_PPT']*0.75, 0)
URBAG_map['03_PPTe'] = np.where(URBAG_map['03_PPT'] >= 4, URBAG_map['03_PPT']*0.75, 0)
URBAG_map['04_PPTe'] = np.where(URBAG_map['04_PPT'] >= 4, URBAG_map['04_PPT']*0.75, 0)
URBAG_map['05_PPTe'] = np.where(URBAG_map['05_PPT'] >= 4, URBAG_map['05_PPT']*0.75, 0)
URBAG_map['06_PPTe'] = np.where(URBAG_map['06_PPT'] >= 4, URBAG_map['06_PPT']*0.75, 0)
URBAG_map['07_PPTe'] = np.where(URBAG_map['07_PPT'] >= 4, URBAG_map['07_PPT']*0.75, 0)
URBAG_map['08_PPTe'] = np.where(URBAG_map['08_PPT'] >= 4, URBAG_map['08_PPT']*0.75, 0)
URBAG_map['09_PPTe'] = np.where(URBAG_map['09_PPT'] >= 4, URBAG_map['09_PPT']*0.75, 0)
URBAG_map['10_PPTe'] = np.where(URBAG_map['10_PPT'] >= 4, URBAG_map['10_PPT']*0.75, 0)
URBAG_map['11_PPTe'] = np.where(URBAG_map['11_PPT'] >= 4, URBAG_map['11_PPT']*0.75, 0)
URBAG_map['12_PPTe'] = np.where(URBAG_map['12_PPT'] >= 4, URBAG_map['12_PPT']*0.75, 0)

pd.unique(URBAG_map['06_PPTe'])

URBAG_map['06_PPT']

## Blue water use / Irrigation demand (ID)

Blue water use (mm)/ Irrigation demand (ID) (mm) per month per plot = CWR – PPTe

URBAG_map["01_ID_1"] = URBAG_map["01_CWR"] - URBAG_map["01_PPTe"]
URBAG_map["02_ID_1"] = URBAG_map["02_CWR"] - URBAG_map["02_PPTe"]
URBAG_map["03_ID_1"] = URBAG_map["03_CWR"] - URBAG_map["03_PPTe"]
URBAG_map["04_ID_1"] = URBAG_map["04_CWR"] - URBAG_map["04_PPTe"]
URBAG_map["05_ID_1"] = URBAG_map["05_CWR"] - URBAG_map["05_PPTe"]
URBAG_map["06_ID_1"] = URBAG_map["06_CWR"] - URBAG_map["06_PPTe"]
URBAG_map["07_ID_1"] = URBAG_map["07_CWR"] - URBAG_map["07_PPTe"]
URBAG_map["08_ID_1"] = URBAG_map["08_CWR"] - URBAG_map["08_PPTe"]
URBAG_map["09_ID_1"] = URBAG_map["09_CWR"] - URBAG_map["09_PPTe"]
URBAG_map["10_ID_1"] = URBAG_map["10_CWR"] - URBAG_map["10_PPTe"]
URBAG_map["11_ID_1"] = URBAG_map["11_CWR"] - URBAG_map["11_PPTe"]
URBAG_map["12_ID_1"] = URBAG_map["12_CWR"] - URBAG_map["12_PPTe"]

np.sort(pd.unique(URBAG_map["01_ID_1"]))

URBAG_map.loc[pd.isna(URBAG_map["01_ID_1"])] ["Fid"]

For irrigation demand that is negative i.e. higher PPTe than CRW we ser the irrigation demando to zero

URBAG_map['01_ID_2'] = np.where(URBAG_map['01_ID_1'] < 0, 0, URBAG_map['01_ID_1'])
URBAG_map['02_ID_2'] = np.where(URBAG_map['02_ID_1'] < 0, 0, URBAG_map['02_ID_1'])
URBAG_map['03_ID_2'] = np.where(URBAG_map['03_ID_1'] < 0, 0, URBAG_map['03_ID_1'])
URBAG_map['04_ID_2'] = np.where(URBAG_map['04_ID_1'] < 0, 0, URBAG_map['04_ID_1'])
URBAG_map['05_ID_2'] = np.where(URBAG_map['05_ID_1'] < 0, 0, URBAG_map['05_ID_1'])
URBAG_map['06_ID_2'] = np.where(URBAG_map['06_ID_1'] < 0, 0, URBAG_map['06_ID_1'])
URBAG_map['07_ID_2'] = np.where(URBAG_map['07_ID_1'] < 0, 0, URBAG_map['07_ID_1'])
URBAG_map['08_ID_2'] = np.where(URBAG_map['08_ID_1'] < 0, 0, URBAG_map['08_ID_1'])
URBAG_map['09_ID_2'] = np.where(URBAG_map['09_ID_1'] < 0, 0, URBAG_map['09_ID_1'])
URBAG_map['10_ID_2'] = np.where(URBAG_map['10_ID_1'] < 0, 0, URBAG_map['10_ID_1'])
URBAG_map['11_ID_2'] = np.where(URBAG_map['11_ID_1'] < 0, 0, URBAG_map['11_ID_1'])
URBAG_map['12_ID_2'] = np.where(URBAG_map['12_ID_1'] < 0, 0, URBAG_map['12_ID_1'])

np.sort(pd.unique(URBAG_map["12_ID_2"]))

Further if catastro says there is no irrigation, then irrigation demand is set to zero

#checking the URBAG map parameter with the cataster information on which plot has (SI) or not (NO) irrigation
URBAG_map['REG_CAT']

URBAG_map['01_ID'] = np.where(URBAG_map['REG_CAT']=="NO", 0, URBAG_map['01_ID_2'])
URBAG_map['02_ID'] = np.where(URBAG_map['REG_CAT']=="NO", 0, URBAG_map['02_ID_2'])
URBAG_map['03_ID'] = np.where(URBAG_map['REG_CAT']=="NO", 0, URBAG_map['03_ID_2'])
URBAG_map['04_ID'] = np.where(URBAG_map['REG_CAT']=="NO", 0, URBAG_map['04_ID_2'])
URBAG_map['05_ID'] = np.where(URBAG_map['REG_CAT']=="NO", 0, URBAG_map['05_ID_2'])
URBAG_map['06_ID'] = np.where(URBAG_map['REG_CAT']=="NO", 0, URBAG_map['06_ID_2'])
URBAG_map['07_ID'] = np.where(URBAG_map['REG_CAT']=="NO", 0, URBAG_map['07_ID_2'])
URBAG_map['08_ID'] = np.where(URBAG_map['REG_CAT']=="NO", 0, URBAG_map['08_ID_2'])
URBAG_map['09_ID'] = np.where(URBAG_map['REG_CAT']=="NO", 0, URBAG_map['09_ID_2'])
URBAG_map['10_ID'] = np.where(URBAG_map['REG_CAT']=="NO", 0, URBAG_map['10_ID_2'])
URBAG_map['11_ID'] = np.where(URBAG_map['REG_CAT']=="NO", 0, URBAG_map['11_ID_2'])
URBAG_map['12_ID'] = np.where(URBAG_map['REG_CAT']=="NO", 0, URBAG_map['12_ID_2'])

np.sort(pd.unique(URBAG_map["12_ID_2"]))

## Water consumed

Water consumed (direct water use) (mm) per month per plot

If cataster says no irrigation then water consumed is set to zero. Else:

Irrigation demand (ID) * (1 + distribution losses (%) + irrigation losses (%))

#distribution efficiency is NAN for some plots which do not need irrigation according to Cataster data
np.sort(pd.unique(URBAG_map['EFIC_DISTR']))

#converting efficiency of distribution to losses due to distribution
#for plots where there is NAN we set the losses to zero
URBAG_map['DISTR_LOSSES'] = np.where(pd.isna(URBAG_map["EFIC_DISTR"]), 0, (100-URBAG_map['EFIC_DISTR']))

#checking there are no NAN in the distribution losses
np.sort(pd.unique(URBAG_map['DISTR_LOSSES']))

#checking there are no NAN in the distribution losses
URBAG_map.loc[pd.isna(URBAG_map["DISTR_LOSSES"])] ["Fid"]

#converting efficiency of irrifation into losses and for plot with no irrigation NAN is set to zero
URBAG_map['RIEGO_LOSSES'] = np.where(pd.isna(URBAG_map['EFIC_RIEGO']), 0, (100-URBAG_map['EFIC_RIEGO']))

#checking the original irrigation efficiency, some plots have NAN
np.sort(pd.unique(URBAG_map['EFIC_RIEGO']))

#checking calculated irrigation losses no NAN should appear
np.sort(pd.unique(URBAG_map['RIEGO_LOSSES']))

#checking calculated irrigation losses no NAN should appear, her the Fid of any plot with NAN is identified
URBAG_map.loc[pd.isna(URBAG_map["RIEGO_LOSSES"])] ["Fid"]

URBAG_map['LOSSES'] = (1 + (URBAG_map['DISTR_LOSSES']/100) + (URBAG_map['RIEGO_LOSSES']/100))

#total losses in irrigation and distribution should be above zero or zero, no NAN should appear
np.sort(pd.unique(URBAG_map['LOSSES']))

#calculating water extraction including the losses of the systems
URBAG_map['01_WE'] = URBAG_map['01_ID'] * URBAG_map['LOSSES']
URBAG_map['02_WE'] = URBAG_map['02_ID'] * URBAG_map['LOSSES']
URBAG_map['03_WE'] = URBAG_map['03_ID'] * URBAG_map['LOSSES'] 
URBAG_map['04_WE'] = URBAG_map['04_ID'] * URBAG_map['LOSSES'] 
URBAG_map['05_WE'] = URBAG_map['05_ID'] * URBAG_map['LOSSES']
URBAG_map['06_WE'] = URBAG_map['06_ID'] * URBAG_map['LOSSES']
URBAG_map['07_WE'] = URBAG_map['07_ID'] * URBAG_map['LOSSES'] 
URBAG_map['08_WE'] = URBAG_map['08_ID'] * URBAG_map['LOSSES'] 
URBAG_map['09_WE'] = URBAG_map['09_ID'] * URBAG_map['LOSSES']
URBAG_map['10_WE'] = URBAG_map['10_ID'] * URBAG_map['LOSSES'] 
URBAG_map['11_WE'] = URBAG_map['11_ID'] * URBAG_map['LOSSES']
URBAG_map['12_WE'] = URBAG_map['12_ID'] * URBAG_map['LOSSES']

URBAG_map.loc[pd.isna(URBAG_map["01_WE"])] ["Fid"]

## OUTPUTS

#Crop water requirement (mm) per month per plot
CWR = pd.DataFrame([URBAG_map['Fid'],
        URBAG_map['Area']*(1/10000),     #Area per plot in Ha       
        URBAG_map['01_CWR'],
        URBAG_map['02_CWR'],
        URBAG_map['03_CWR'],
        URBAG_map['04_CWR'],
        URBAG_map['05_CWR'],
        URBAG_map['06_CWR'],
        URBAG_map['07_CWR'],
        URBAG_map['08_CWR'],
        URBAG_map['09_CWR'],
        URBAG_map['10_CWR'],
        URBAG_map['11_CWR'],
        URBAG_map['12_CWR']]).T

CWR

sns.set()

#mm per month
df=CWR.drop(['Fid',"Area"], axis=1)
sns.boxplot(data=df, orient="h")
plt.xlabel('mm per month')
plt.title('Crop Water Requirement (CWR)')

#Green water use (mm) per month per plot
Green_water = pd.DataFrame([URBAG_map['Fid'],
        URBAG_map['Area']*(1/10000),     #Area per plot in Ha
        URBAG_map['01_PPTe'],
        URBAG_map['02_PPTe'],
        URBAG_map['03_PPTe'],
        URBAG_map['04_PPTe'],
        URBAG_map['05_PPTe'],
        URBAG_map['06_PPTe'],
        URBAG_map['07_PPTe'],
        URBAG_map['08_PPTe'],
        URBAG_map['09_PPTe'],
        URBAG_map['10_PPTe'],
        URBAG_map['11_PPTe'],
        URBAG_map['12_PPTe']]).T

Green_water

#mm per month
df2 = Green_water.drop(['Fid',"Area"], axis=1)
sns.boxplot(data=df2, orient="h")
plt.xlabel('mm per month')
plt.title('Green Water')

#Blue water use/Irrigation demand (mm) per month per plot
Irrigation_Demand = pd.DataFrame([URBAG_map['Fid'],
                                URBAG_map['Area']*(1/10000),     #Area per plot in Ha
                                URBAG_map['01_ID'],
                                URBAG_map['02_ID'],
                                URBAG_map['03_ID'],
                                URBAG_map['04_ID'],
                                URBAG_map['05_ID'],
                                URBAG_map['06_ID'],
                                URBAG_map['07_ID'],
                                URBAG_map['08_ID'],
                                URBAG_map['09_ID'],
                                URBAG_map['10_ID'],
                                URBAG_map['11_ID'],
                                URBAG_map['12_ID']]).T

Irrigation_Demand

#mm per month
df3 = Irrigation_Demand.drop(['Fid',"Area"], axis=1)
sns.boxplot(data=df3, orient="h")
plt.xlabel('mm per month')
plt.title('Irrigation Demand')

#Direct water use/Irrigation (mm) per month per plot
Water_extraction = pd.DataFrame([URBAG_map['Fid'],
                            URBAG_map['Area']*(1/10000),     #Area per plot in Ha
                            URBAG_map['01_WE'],
                            URBAG_map['02_WE'],
                            URBAG_map['03_WE'],
                            URBAG_map['04_WE'],
                            URBAG_map['05_WE'],
                            URBAG_map['06_WE'],
                            URBAG_map['07_WE'],
                            URBAG_map['08_WE'],
                            URBAG_map['09_WE'],
                            URBAG_map['10_WE'],
                            URBAG_map['11_WE'],
                            URBAG_map['12_WE']]).T

Water_extraction

Water_extraction.keys()

#mm per month
df4 = Water_extraction.drop(['Fid',"Area"], axis=1)
sns.boxplot(data=df4, orient="h")
plt.xlabel('mm per month')
plt.title('Water Extraction')

## Conversion to m3/ha yr

### Irrigation demand conversion from mm/ha * month to m3/ha * yr

#First step is to find the mm/month per ha! in order to convert later to m3
Irrigation_Demand_ha_month = pd.DataFrame()
Irrigation_Demand_ha_month['01_ID'] = Irrigation_Demand['01_ID']/Irrigation_Demand['Area']
Irrigation_Demand_ha_month['02_ID'] = Irrigation_Demand['02_ID']/Irrigation_Demand['Area']
Irrigation_Demand_ha_month['03_ID'] = Irrigation_Demand['03_ID']/Irrigation_Demand['Area']
Irrigation_Demand_ha_month['04_ID'] = Irrigation_Demand['04_ID']/Irrigation_Demand['Area']
Irrigation_Demand_ha_month['05_ID'] = Irrigation_Demand['05_ID']/Irrigation_Demand['Area']
Irrigation_Demand_ha_month['06_ID'] = Irrigation_Demand['06_ID']/Irrigation_Demand['Area']
Irrigation_Demand_ha_month['07_ID'] = Irrigation_Demand['07_ID']/Irrigation_Demand['Area']
Irrigation_Demand_ha_month['08_ID'] = Irrigation_Demand['08_ID']/Irrigation_Demand['Area']
Irrigation_Demand_ha_month['09_ID'] = Irrigation_Demand['09_ID']/Irrigation_Demand['Area']
Irrigation_Demand_ha_month['10_ID'] = Irrigation_Demand['10_ID']/Irrigation_Demand['Area']
Irrigation_Demand_ha_month['11_ID'] = Irrigation_Demand['11_ID']/Irrigation_Demand['Area']
Irrigation_Demand_ha_month['12_ID'] = Irrigation_Demand['12_ID']/Irrigation_Demand['Area']

Irrigation_Demand_ha_month

col_list_ID= list(Irrigation_Demand_ha_month)
col_list_ID

ID_ha_yr = pd.DataFrame()
ID_ha_yr ["Fid"] = Irrigation_Demand["Fid"]
ID_ha_yr ["Area"] = Irrigation_Demand["Area"] #Area in Ha
ID_ha_yr ["ID_mm_ha_yr"] = Irrigation_Demand_ha_month[col_list_ID].sum(axis=1) #mm per ha year
ID_ha_yr

ID_m3_ha_yr = pd.DataFrame()
ID_m3_ha_yr ["Fid"] = ID_ha_yr["Fid"]
ID_m3_ha_yr ["Area"] = ID_ha_yr["Area"] #Area in Ha
ID_m3_ha_yr ["ID_m3_ha_yr"] = ID_ha_yr["ID_mm_ha_yr"] * 10 #m3 per ha year
ID_m3_ha_yr

### Water_extraction conversion from mm/ha * month to m3/ha * yr

#First step is to find the mm/month per ha! in order to convert later to m3
WE_ha_month = pd.DataFrame()
WE_ha_month["Fid"]   = Water_extraction['Fid']
WE_ha_month['01_WE'] = Water_extraction['01_WE']/Water_extraction['Area']
WE_ha_month['02_WE'] = Water_extraction['02_WE']/Water_extraction['Area']
WE_ha_month['03_WE'] = Water_extraction['03_WE']/Water_extraction['Area']
WE_ha_month['04_WE'] = Water_extraction['04_WE']/Water_extraction['Area']
WE_ha_month['05_WE'] = Water_extraction['05_WE']/Water_extraction['Area']
WE_ha_month['06_WE'] = Water_extraction['06_WE']/Water_extraction['Area']
WE_ha_month['07_WE'] = Water_extraction['07_WE']/Water_extraction['Area']
WE_ha_month['08_WE'] = Water_extraction['08_WE']/Water_extraction['Area']
WE_ha_month['09_WE'] = Water_extraction['09_WE']/Water_extraction['Area']
WE_ha_month['10_WE'] = Water_extraction['10_WE']/Water_extraction['Area']
WE_ha_month['11_WE'] = Water_extraction['11_WE']/Water_extraction['Area']
WE_ha_month['12_WE'] = Water_extraction['12_WE']/Water_extraction['Area']

WE_ha_month

WE_ha_month.loc[pd.isna(WE_ha_month["01_WE"])] ["Fid"]

Water_extraction.loc[WE_ha_month["Fid"]==559]["Area"]

WE_ha_month.loc[WE_ha_month["Fid"]==559]

#mm/ha per month
df5 = WE_ha_month.drop(['Fid'], axis=1)
sns.boxplot(data=df5, orient="h", showfliers=False)
plt.xlabel('mm per month')
plt.title('Water extraction')

col_list= list(WE_ha_month)
col_list.remove('Fid')
col_list

WE_ha_yr = pd.DataFrame()
WE_ha_yr ["Fid"] = Water_extraction["Fid"]
WE_ha_yr ["Area"] = Water_extraction["Area"] #Area in Ha
WE_ha_yr ["WE_mm_ha_yr"] = WE_ha_month[col_list].sum(axis=1) #mm per ha year
WE_ha_yr

WE_ha_yr["WE_mm_ha_yr"].sort_values(ascending=True)

WE_ha_yr["WE_mm_ha_yr"].plot(kind="box", showfliers=False)
plt.title('Water extraction - mm per ha yr')

WE_m3_ha_yr = pd.DataFrame()
WE_m3_ha_yr ["Fid"] = WE_ha_yr["Fid"]
WE_m3_ha_yr ["Area"] = WE_ha_yr["Area"] #Area in Ha
WE_m3_ha_yr ["WE_m3_ha_yr"] = WE_ha_yr["WE_mm_ha_yr"] * 10 #m3 per ha year
WE_m3_ha_yr

WE_m3_ha_yr ["WE_m3_ha_yr"].sort_values(ascending=True)

URBAG_map["WE_m3_ha_yr"] = WE_m3_ha_yr ["WE_m3_ha_yr"]

URBAG_map["WE_m3_ha_yr"].plot(kind="box", showfliers=False)
plt.title('Water extraction - m3 per ha yr')

URBAG_map = URBAG_map.drop(['01_ID_1', '02_ID_1',
       '03_ID_1', '04_ID_1', '05_ID_1', '06_ID_1', '07_ID_1', '08_ID_1',
       '09_ID_1', '10_ID_1', '11_ID_1', '12_ID_1', '01_ID_2', '02_ID_2',
       '03_ID_2', '04_ID_2', '05_ID_2', '06_ID_2', '07_ID_2', '08_ID_2',
       '09_ID_2', '10_ID_2', '11_ID_2', '12_ID_2', 'DISTR_LOSSES', 'RIEGO_LOSSES'], axis=1)

URBAG_map.keys().values

#1e6 m3 = hm3
AMB_total_water_irrigation_demand = np.sum(ID_m3_ha_yr["Area"]*ID_m3_ha_yr["ID_m3_ha_yr"])/1e6

print("Total water irrigation demand in hm3/yr is "+ str(AMB_total_water_irrigation_demand))

#1e6 m3 = hm3
AMB_total_water_extraction = np.sum((URBAG_map["Area"]*(1/10000))*URBAG_map["WE_m3_ha_yr"])/1e6

print("Total water extraction in hm3/yr is "+ str(AMB_total_water_extraction))

## Saving outputs

%store URBAG_map

CWR.to_csv("output/" + scen + "/water/CWR.csv") #mm/month crop water requirements
Green_water.to_csv("output/" + scen + "/water/Green_water.csv") #mm/month Effective precipitation
Irrigation_Demand.to_csv("output/" + scen + "/water/Irrigation_Demand.csv") #mm/month
Water_extraction.to_csv("output/" + scen + "/water/Water_extraction.csv") #mm/month after losses
WE_ha_yr.to_csv("output/" + scen + "/water/WE_ha_yr.csv") #mm /ha * year
WE_m3_ha_yr.to_csv("output/" + scen + "/water/WE_m3_ha_yr.csv") #m3 /ha * year
