==> 7_Total_LCIA.ipynb <==
# Total LCIA

This notebook calculates the total LCIA as the sum of direct and indirect LCIA

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns

#In this cell the scenario is chosen. This can either be defined here or in the RUN_ALL.ipynb
#If defined here unmark the scen = line and select the scenario you will run.
#If defined in RUN_ALL.ipynb unmark the %store -r scen line.
#Scenario names are:
#S0_MinFert > N,P,K demand met by mineral fertilizers only
#S0_struvite_P > P demand met by struvite, N demand met by struvite and mineral fertilizer and K demand met by mineral fertilizer
#S0_compost > N,P,K supplied by compost produced in the AMB, remaining N,P,K demand met by mineral fertilizer
#S0_Ammon_salts > N supplied from recovered ammonium salts and from mineral fertilizers, P from struvite and K from mineral fertilizer
scen = 'S0_MinFert'
#%store -r scen

print(scen)

%store -r data_back
%store -r data_foreground_1

## Background impacts

data_back.keys()

background_impacts = data_back.copy()
background_impacts

background_impacts.keys()

pd.unique(background_impacts['database'])

## Foreground impacts

data_foreground_1.keys()

foreground_impacts = data_foreground_1.copy()

foreground_impacts

## Total LCIA

#concatenating both background and extended foreground
data_all_LCIA = pd.concat([background_impacts,foreground_impacts],ignore_index=True)

#making a copy create the total impacts in another dataframe
total_impacts = data_all_LCIA.copy()

total_impacts.keys()

pd.unique(total_impacts['database'])

pd.unique(total_impacts['product'])

pd.unique(total_impacts['imp_cat'])

pd.unique(total_impacts['Scenario'])

total_impacts

total_impacts_GW = total_impacts[(total_impacts['imp_cat']=='GWP100')]
total_impacts_ME = total_impacts[(total_impacts['imp_cat']=='ME')]
total_impacts_ME_Reg = total_impacts[(total_impacts['imp_cat']=='ME_Reg')]
total_impacts_FE = total_impacts[(total_impacts['imp_cat']=='FE')]
total_impacts_FE_Reg = total_impacts[(total_impacts['imp_cat']=='FE_Reg')]
total_impacts_WC = total_impacts[(total_impacts['imp_cat']=='WC')]
total_impacts_ADP = total_impacts[(total_impacts['imp_cat']=='ADP')]

print(total_impacts_GW.loc[total_impacts_GW["database"]=="ei391"])

print(total_impacts_ME.loc[total_impacts_ME["database"]=="ei391"])

print(total_impacts_ME_Reg.loc[total_impacts_ME_Reg["database"]=="ei391"])

print(total_impacts_FE.loc[total_impacts_FE["database"]=="ei391"])

print(total_impacts_FE_Reg.loc[total_impacts_FE_Reg["database"]=="ei391"])

print(total_impacts_ADP.loc[total_impacts_ADP["database"]=="ei391"])

print(total_impacts_WC.loc[total_impacts_WC["database"]=="ei391"])

## Saving outputs

%store total_impacts
%store total_impacts_GW
%store total_impacts_ME
%store total_impacts_ME_Reg
%store total_impacts_FE
%store total_impacts_FE_Reg
%store total_impacts_WC

#Total_impacts + foreground_impacts
total_impacts.to_csv("output/" + scen + "/LCIA/Total_LCIA.csv")
total_impacts_GW.to_csv("output/" + scen + "/LCIA/Total_LCIA_GWP.csv")
total_impacts_ME.to_csv("output/" + scen + "/LCIA/Total_LCIA_ME.csv")
total_impacts_ME_Reg.to_csv("output/" + scen + "/LCIA/Total_LCIA_ME_Reg.csv")
total_impacts_FE.to_csv("output/" + scen + "/LCIA/Total_LCIA_FE.csv")
total_impacts_FE_Reg.to_csv("output/" + scen + "/LCIA/Total_LCIA_FE_Reg.csv")
total_impacts_WC.to_csv("output/" + scen + "/LCIA/Total_LCIA_WC.csv")
