==> 8_1_N_P_Balances_allscenarios.ipynb <==
# Vizualization of balanced inputs for all scenarios

This notebook picks up the results from the balances inputs and plots each nutrient in one plot.

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns

N_balance_S0_compost = pd.read_csv("output/S0_compost/balances/Nitrogen_Balance_new.csv")
N_balance_S0_MinFert = pd.read_csv("output/S0_MinFert/balances/Nitrogen_Balance_new.csv")
N_balance_S0_struvite_P = pd.read_csv("output/S0_struvite_P/balances/Nitrogen_Balance_new.csv")
N_balance_S0_Ammon_salts = pd.read_csv("output/S0_Ammon_salts/balances/Nitrogen_Balance_new.csv")

P_balance_S0_compost = pd.read_csv("output/S0_compost/balances/Phosphorus_Balance_new.csv")
P_balance_S0_MinFert = pd.read_csv("output/S0_MinFert/balances/Phosphorus_Balance_new.csv")
P_balance_S0_struvite_P = pd.read_csv("output/S0_struvite_P/balances/Phosphorus_Balance_new.csv")
P_balance_S0_Ammon_salts = pd.read_csv("output/S0_Ammon_salts/balances/Phosphorus_Balance_new.csv")

K_balance_S0_compost = pd.read_csv("output/S0_compost/balances/Potassium_Balance_new.csv")
K_balance_S0_MinFert = pd.read_csv("output/S0_MinFert/balances/Potassium_Balance_new.csv")
K_balance_S0_struvite_P = pd.read_csv("output/S0_struvite_P/balances/Potassium_Balance_new.csv")
K_balance_S0_Ammon_salts = pd.read_csv("output/S0_Ammon_salts/balances/Potassium_Balance_new.csv")

N_balance_S0_compost["Scenario"]    = "compost"
N_balance_S0_MinFert["Scenario"]    = "MinFert"
N_balance_S0_struvite_P["Scenario"] = "struvite_P"
N_balance_S0_Ammon_salts["Scenario"] = "Ammon_salts"
P_balance_S0_compost["Scenario"]    = "compost"
P_balance_S0_MinFert["Scenario"]    = "MinFert"
P_balance_S0_struvite_P["Scenario"] = "struvite_P"
P_balance_S0_Ammon_salts["Scenario"] = "Ammon_salts"
K_balance_S0_compost["Scenario"]    = "compost"
K_balance_S0_MinFert["Scenario"]    = "MinFert"
K_balance_S0_struvite_P["Scenario"] = "struvite_P"
K_balance_S0_Ammon_salts["Scenario"] = "Ammon_salts"

N_pdList = [N_balance_S0_compost, N_balance_S0_MinFert, N_balance_S0_struvite_P,N_balance_S0_Ammon_salts]  # List of your dataframes
N_balance_all_scenarios = pd.concat(N_pdList)

P_pdList = [P_balance_S0_compost, P_balance_S0_MinFert, P_balance_S0_struvite_P,P_balance_S0_Ammon_salts]  # List of your dataframes
P_balance_all_scenarios = pd.concat(P_pdList)

K_pdList = [K_balance_S0_compost, K_balance_S0_MinFert, K_balance_S0_struvite_P,K_balance_S0_Ammon_salts]  # List of your dataframes
K_balance_all_scenarios = pd.concat(K_pdList)

fig, axs = plt.subplots(4, 4, figsize=(17,12))
N_balance_all_scenarios.boxplot(column="N_Manure", by='Scenario', ax=axs[0,0], showmeans=True, grid=False, showfliers=False)
axs[0, 0].set_title("Manure")
N_balance_all_scenarios.boxplot(column="N_Fertiliser", by='Scenario', ax=axs[0,1], showmeans=True, grid=False, showfliers=False)
axs[0, 1].set_title("Mineral Fertiliser")
N_balance_all_scenarios.boxplot(column="N_Ammonium_salts", by='Scenario', ax=axs[0,2], showmeans=True, grid=False, showfliers=False)
axs[0, 2].set_title("Recovered ammonium salts")
N_balance_all_scenarios.boxplot(column="N_Residues", by='Scenario', ax=axs[0,3], showmeans=True, grid=False, showfliers=False)
axs[0, 3].set_title("Agricultural Residues")
N_balance_all_scenarios.boxplot(column="N_Struvite", by='Scenario', ax=axs[1,0], grid=False, showfliers=False)
axs[1, 0].set_title("Struvite")
N_balance_all_scenarios.boxplot(column="N_Compost", by='Scenario', ax=axs[1,1], grid=False, showfliers=False)
axs[1, 1].set_title("Compost")
N_balance_all_scenarios.boxplot(column="N_Fixation", by='Scenario', ax=axs[1,2], showmeans=True, grid=False, showfliers=False)
axs[1, 2].set_title("N Fixation")
N_balance_all_scenarios.boxplot(column="N_in_incl_N_fix", by='Scenario', ax=axs[1,3], showmeans=True, grid=False, showfliers=False)
axs[1, 3].set_title("Total N inputs incl. fixation")
N_balance_all_scenarios.boxplot(column="N_NH3", by='Scenario', ax=axs[2,0], showmeans=True, grid=False, showfliers=False)
axs[2, 0].set_title("NH$_{3}$ emissions")
N_balance_all_scenarios.boxplot(column="N_NOx", by='Scenario', ax=axs[2,1], showmeans=True, grid=False, showfliers=False)
axs[2, 1].set_title("NO$_{x}$ emissions")
N_balance_all_scenarios.boxplot(column="N_N2O", by='Scenario', ax=axs[2,2], showmeans=True, grid=False, showfliers=False)
axs[2, 2].set_title("N$_{2}$O emissions")
N_balance_all_scenarios.boxplot(column="N_NO3", by='Scenario', ax=axs[2,3], showmeans=True, grid=False, showfliers=False)
axs[2, 3].set_title("NO$_{3}^-$ emissions")
N_balance_all_scenarios.boxplot(column="N_Harvest", by='Scenario', ax=axs[3,0], showmeans=True, grid=False, showfliers=False)
axs[3, 0].set_title("N removed with Harvest")
N_balance_all_scenarios.boxplot(column="N_Out", by='Scenario', ax=axs[3,1], showmeans=True, grid=False, showfliers=False)
axs[3, 1].set_title("Total N output")
fig.delaxes(axs[3,2])
fig.delaxes(axs[3,3])
plt.tight_layout()
fig.suptitle("N inputs and outputs - kg N/ha yr")
fig.savefig( 'output/N_balance_all_scenarios' + '.png')

fig, axs = plt.subplots(2, 4, figsize=(15,6))
P_balance_all_scenarios.boxplot(column="P_Manure", by='Scenario', ax=axs[0,0], showmeans=True, grid=False, showfliers=False)
axs[0, 0].set_title("Manure")
P_balance_all_scenarios.boxplot(column="P_Fertiliser", by='Scenario', ax=axs[0,1], showmeans=True, grid=False, showfliers=False)
axs[0, 1].set_title("Mineral Fertiliser")
P_balance_all_scenarios.boxplot(column="P_Struvite", by='Scenario', ax=axs[0,2], showmeans=True, grid=False, showfliers=False)
axs[0, 2].set_title("Struvite")
P_balance_all_scenarios.boxplot(column="P_Compost", by='Scenario', ax=axs[0,3], showmeans=True, grid=False, showfliers=False)
axs[0, 3].set_title("Compost")

P_balance_all_scenarios.boxplot(column="P_In", by='Scenario', ax=axs[1,0], showmeans=True, grid=False, showfliers=False)
axs[1, 0].set_title("Total P inputs")
P_balance_all_scenarios.boxplot(column="P_PO43-", by='Scenario', ax=axs[1,1], showmeans=True, grid=False, showfliers=False)
axs[1, 1].set_title("PO$_{4}^3-$ runoff")
P_balance_all_scenarios.boxplot(column="P_Harvest", by='Scenario', ax=axs[1,2], showmeans=True, grid=False, showfliers=False)
axs[1, 2].set_title("P removed \n with harvest")
P_balance_all_scenarios.boxplot(column="P_Out", by='Scenario', ax=axs[1,3], showmeans=True, grid=False, showfliers=False)
axs[1, 3].set_title("Total P output")
plt.tight_layout()
fig.suptitle("P inputs and outputs - kg P2O5/ha yr")
fig.savefig( 'output/P_balance_all_scenarios' + '.png')

fig, axs = plt.subplots(1, 4, figsize=(15,4))
K_balance_all_scenarios.boxplot(column="K_Fertiliser", by='Scenario', ax=axs[0], showmeans=True, grid=False, showfliers=False)
axs[0].set_title("Mineral Fertilizer")
K_balance_all_scenarios.boxplot(column="K_Compost", by='Scenario', ax=axs[1], showmeans=True, grid=False, showfliers=False)
axs[1].set_title("Compost")
K_balance_all_scenarios.boxplot(column="K_In", by='Scenario', ax=axs[2], showmeans=True, grid=False, showfliers=False)
axs[2].set_title("Total inputs")
K_balance_all_scenarios.boxplot(column="K_Harv", by='Scenario', ax=axs[3], showmeans=True, grid=False, showfliers=False)
axs[3].set_title("K removed with harvest")
plt.tight_layout()
fig.suptitle("K inputs and outputs - kg K2O/ha yr")
fig.savefig( 'output/K_balance_all_scenarios' + '.png')
