==> 8_2_Total_scenarios.ipynb <==
# Figures Multiple Scenarios

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
import plotnine as p9

%store -r Totals_AMB

LCIA_S0_MinFert = pd.read_csv("output/S0_MinFert/LCIA/Total_LCIA.csv")
LCIA_S0_MinFert = LCIA_S0_MinFert.drop("Unnamed: 0", axis=1)

LCIA_S0_struvite_P = pd.read_csv("output/S0_struvite_P/LCIA/Total_LCIA.csv")
LCIA_S0_struvite_P = LCIA_S0_struvite_P.drop("Unnamed: 0", axis=1)

LCIA_S0_Compost = pd.read_csv("output/S0_compost/LCIA/Total_LCIA.csv")
LCIA_S0_Compost = LCIA_S0_Compost.drop('Unnamed: 0', axis=1)

LCIA_S0_Ammon_salts = pd.read_csv("output/S0_Ammon_salts/LCIA/Total_LCIA.csv")
LCIA_S0_Ammon_salts = LCIA_S0_Ammon_salts.drop('Unnamed: 0', axis=1)

All_scenarios_LCIA = pd.concat([LCIA_S0_MinFert,LCIA_S0_struvite_P, LCIA_S0_Compost, LCIA_S0_Ammon_salts],ignore_index=True)
All_scenarios_LCIA

All_scenarios_LCIA.keys()

pd.unique(All_scenarios_LCIA['database'])

test = All_scenarios_LCIA[All_scenarios_LCIA["imp_cat"]=="GWP100"]

test2= test[test["database"]=="ei391" ]

test2[test2["Scenario"]=="S0_MinFert"]

GW_LCIA = All_scenarios_LCIA[(All_scenarios_LCIA['imp_cat']=='GWP100')]
ME_LCIA = All_scenarios_LCIA[(All_scenarios_LCIA['imp_cat']=='ME')]
ME_Reg_LCIA = All_scenarios_LCIA[(All_scenarios_LCIA['imp_cat']=='ME_Reg')]
FE_LCIA = All_scenarios_LCIA[(All_scenarios_LCIA['imp_cat']=='FE')]
FE_Reg_LCIA = All_scenarios_LCIA[(All_scenarios_LCIA['imp_cat']=='FE_Reg')]
WC_LCIA = All_scenarios_LCIA[(All_scenarios_LCIA['imp_cat']=='WC')]
ADP_LCIA = All_scenarios_LCIA[(All_scenarios_LCIA['imp_cat']=='ADP')]

GW_LCIA.keys()

All_scenarios_LCIA.keys()

Totals_contribution = All_scenarios_LCIA.groupby(["imp_cat",'Scenario','database',"product"]).sum("imp_yr").reset_index()
Totals_contribution

Totals = All_scenarios_LCIA.groupby(["imp_cat",'Scenario','database']).sum("imp_yr").reset_index()
Totals["Area_HA"] = Totals_AMB[0]["Area_HA"]
Totals["FoodProd_tonsYR"] = Totals_AMB[0]["Food_tonsYR"]

Totals_contribution.to_csv("output/Totals_contribution.csv")

Totals.to_csv("output/Totals.csv")

Totals.loc[(Totals["imp_cat"]=="GWP100") & (Totals["database"]=="ei391")]

# Visualizations

#Climate change
fig = plt.figure()
plot = (p9.ggplot(GW_LCIA, p9.aes('database','imp_yr', fill='product')) +\
            p9.geom_col() +\
            #p9.scale_x_discrete(limits=total_impacts_GW['database'].tolist()) +\
            #p9.facet_grid('LCIA ~ Scenario') +\
            p9.scale_fill_manual(values={'CH4':"#f2d8dd",
                                        #'CO':"#FF9999",
                                        'CO2':"#f5a9b8",
                                        'N2O':"#FF0033",
                                        #'NOx':"#FF0000",
                                        'Irrigation':"#b3495e",
                                        'Nitrate':"#7a1c2f",
                                        'Phosphate':"#590919",
                                        'inorganic nitrogen fertiliser, as N':"#6666FF",
                                        'inorganic phosphorus fertiliser, as P2O5':"#0000FF",
                                        'inorganic potassium fertiliser, as K2O':"#000099",
                                        'Struvite at WWTP':"#CCCCFF",
                                        'Ammonium sulphate at WWTP':"#f0f0fc",
                                        'compost, market for compost Metropolitan Area of Barcelona':"#5ec7fa"}, name=' ') +\
            p9.facet_wrap("Scenario") +\
            p9.labels.ylab('Climate Change (Tons CO$_{2eq}$ yr$^{-1}$)') +\
            p9.labels.xlab('') +\
            p9.theme(figure_size=(8, 6), axis_text_x=p9.element_text(rotation=90, hjust = 1))).draw(show=True)
fig.savefig('output/GW_LCIA_yr_allscenarios.png', bbox_inches='tight')

#Marine Eutrophication
fig = plt.figure()
plot = (p9.ggplot(ME_LCIA, p9.aes('database','imp_yr', fill='product')) +\
            p9.geom_col() +\
            #p9.scale_x_discrete(limits=total_impacts_GW['database'].tolist()) +\
            p9.scale_fill_manual(values={'CH4':"#f2d8dd",
                                        #'CO':"#FF9999",
                                        'CO2':"#f5a9b8",
                                        'N2O':"#FF0033",
                                        #'NOx':"#FF0000",
                                        'Irrigation':"#b3495e",
                                        'Nitrate':"#7a1c2f",
                                        'Phosphate':"#590919",
                                        'inorganic nitrogen fertiliser, as N':"#6666FF",
                                        'inorganic phosphorus fertiliser, as P2O5':"#0000FF",
                                        'inorganic potassium fertiliser, as K2O':"#000099",
                                        'Struvite at WWTP':"#CCCCFF",
                                        'Ammonium sulphate at WWTP':"#f0f0fc",
                                        'compost, market for compost Metropolitan Area of Barcelona':"#5ec7fa"}, name=' ') +\
           #p9.facet_grid('Scenario ~ LCIA') +\
            p9.facet_wrap(["Scenario"]) +\
            p9.labels.ylab('Marine Eutrophication (Tons N$_{eq}$ yr$^{-1}$)') +\
            p9.labels.xlab('') +\
            p9.theme(figure_size=(8, 6),  axis_text_x=p9.element_text(rotation=90, hjust = 1))).draw(show=True)
fig.savefig('output/ME_LCIA_yr_allscenarios.png', bbox_inches='tight')

#Marine Eutrophication Regionalized 
fig = plt.figure()
plot = (p9.ggplot(ME_Reg_LCIA, p9.aes('database','imp_yr', fill='product')) +\
            p9.geom_col() +\
            #p9.scale_x_discrete(limits=total_impacts_GW['database'].tolist()) +\
            p9.scale_fill_manual(values={'CH4':"#f2d8dd",
                                        #'CO':"#FF9999",
                                        'CO2':"#f5a9b8",
                                        'N2O':"#FF0033",
                                        #'NOx':"#FF0000",
                                        'Irrigation':"#b3495e",
                                        'Nitrate':"#7a1c2f",
                                        'Phosphate':"#590919",
                                        'inorganic nitrogen fertiliser, as N':"#6666FF",
                                        'inorganic phosphorus fertiliser, as P2O5':"#0000FF",
                                        'inorganic potassium fertiliser, as K2O':"#000099",
                                        'Struvite at WWTP':"#CCCCFF",
                                        'Ammonium sulphate at WWTP':"#f0f0fc",
                                        'compost, market for compost Metropolitan Area of Barcelona':"#5ec7fa"}, name=' ') +\
            #p9.facet_grid('Scenario ~ LCIA') +\
            p9.facet_wrap(["Scenario"]) +\
            p9.labels.ylab('Regionalized Marine Eutrophication (Tons N$_{eq}$ yr$^{-1}$)') +\
            p9.labels.xlab('') +\
            p9.theme(figure_size=(8, 6), axis_text_x=p9.element_text(rotation=90, hjust = 1))).draw(show=True)
fig.savefig('output/ME_Reg_LCIA_yr_allscenarios.png', bbox_inches='tight')

#Freshwater Eutrophication
fig = plt.figure()
plot = (p9.ggplot(FE_LCIA, p9.aes('database','imp_yr', fill='product')) +\
            p9.geom_col() +\
            #p9.scale_x_discrete(limits=total_impacts_GW['database'].tolist()) +\
            p9.scale_fill_manual(values={'CH4':"#f2d8dd",
                                        #'CO':"#FF9999",
                                        'CO2':"#f5a9b8",
                                        'N2O':"#FF0033",
                                        #'NOx':"#FF0000",
                                        'Irrigation':"#b3495e",
                                        'Nitrate':"#7a1c2f",
                                        'Phosphate':"#590919",
                                        'inorganic nitrogen fertiliser, as N':"#6666FF",
                                        'inorganic phosphorus fertiliser, as P2O5':"#0000FF",
                                        'inorganic potassium fertiliser, as K2O':"#000099",
                                        'Struvite at WWTP':"#CCCCFF",
                                        'Ammonium sulphate at WWTP':"#f0f0fc",
                                        'compost, market for compost Metropolitan Area of Barcelona':"#5ec7fa"}, name=' ') +\
            #p9.facet_grid('Scenario ~ LCIA') +\
            p9.facet_wrap(["Scenario"]) +\
            p9.labels.ylab('Freshwater Eutrophication (Tons P$_{eq}$ yr$^{-1}$)') +\
            p9.labels.xlab('') +\
            p9.theme(figure_size=(8, 6), axis_text_x=p9.element_text(rotation=90, hjust = 1))).draw(show=True)
fig.savefig('output/FE_LCIA_yr_allscenarios.png', bbox_inches='tight')

#Freshwater Eutrophication Regionalized foreground impacts
fig = plt.figure()
plot = (p9.ggplot(FE_Reg_LCIA, p9.aes('database','imp_yr', fill='product')) +\
            p9.geom_col() +\
            #p9.scale_x_discrete(limits=total_impacts_GW['database'].tolist()) +\
            p9.scale_fill_manual(values={'CH4':"#f2d8dd",
                                        #'CO':"#FF9999",
                                        'CO2':"#f5a9b8",
                                        'N2O':"#FF0033",
                                        #'NOx':"#FF0000",
                                        'Irrigation':"#b3495e",
                                        'Nitrate':"#7a1c2f",
                                        'Phosphate':"#590919",
                                        'inorganic nitrogen fertiliser, as N':"#6666FF",
                                        'inorganic phosphorus fertiliser, as P2O5':"#0000FF",
                                        'inorganic potassium fertiliser, as K2O':"#000099",
                                        'Struvite at WWTP':"#CCCCFF",
                                        'Ammonium sulphate at WWTP':"#f0f0fc",
                                        'compost, market for compost Metropolitan Area of Barcelona':"#5ec7fa"}, name=' ') +\
            #p9.facet_grid('Scenario ~ LCIA') +\
            p9.facet_wrap(["Scenario"]) +\
            p9.labels.ylab('Regionalized Freshwater Eutrophication (Tons P$_{eq}$ yr$^{-1}$)') +\
            p9.labels.xlab('') +\
            p9.theme(figure_size=(8, 6), axis_text_x=p9.element_text(rotation=90, hjust = 1))).draw(show=True)
fig.savefig('output/FE_Reg_LCIA_yr_allscenarios.png', bbox_inches='tight')

#Water depletion
fig = plt.figure()
plot = (p9.ggplot(WC_LCIA, p9.aes('database','imp_yr', fill='product')) +\
            p9.geom_col() +\
            #p9.scale_x_discrete(limits=total_impacts_GW['database'].tolist()) +\
            p9.scale_fill_manual(values={'CH4':"#f2d8dd",
                                        #'CO':"#FF9999",
                                        'CO2':"#f5a9b8",
                                        'N2O':"#FF0033",
                                        #'NOx':"#FF0000",
                                        'Irrigation':"#b3495e",
                                        'Nitrate':"#7a1c2f",
                                        'Phosphate':"#590919",
                                        'inorganic nitrogen fertiliser, as N':"#6666FF",
                                        'inorganic phosphorus fertiliser, as P2O5':"#0000FF",
                                        'inorganic potassium fertiliser, as K2O':"#000099",
                                        'Struvite at WWTP':"#CCCCFF",
                                        'Ammonium sulphate at WWTP':"#f0f0fc",
                                        'compost, market for compost Metropolitan Area of Barcelona':"#5ec7fa"}, name=' ') +\
            #p9.facet_grid('Scenario ~ LCIA') +\
            p9.facet_wrap(["Scenario"]) +\
            p9.labels.ylab('Water Consumption (*1000 m$^{3}$ yr$^{-1}$)') +\
            p9.labels.xlab('') +\
            p9.theme(figure_size=(8, 6), axis_text_x=p9.element_text(rotation=90, hjust = 1))).draw(show=True)
fig.savefig('output/WD_LCIA_yr_allscenarios.png', bbox_inches='tight')

#Abioditc resource depletion
fig = plt.figure()
plot = (p9.ggplot(ADP_LCIA, p9.aes('database','imp_yr', fill='product')) +\
            p9.geom_col() +\
            #p9.scale_x_discrete(limits=total_impacts_GW['database'].tolist()) +\
            p9.scale_fill_manual(values={'CH4':"#f2d8dd",
                                        #'CO':"#FF9999",
                                        'CO2':"#f5a9b8",
                                        'N2O':"#FF0033",
                                        #'NOx':"#FF0000",
                                        'Irrigation':"#b3495e",
                                        'Nitrate':"#7a1c2f",
                                        'Phosphate':"#590919",
                                        'inorganic nitrogen fertiliser, as N':"#6666FF",
                                        'inorganic phosphorus fertiliser, as P2O5':"#0000FF",
                                        'inorganic potassium fertiliser, as K2O':"#000099",
                                        'Struvite at WWTP':"#CCCCFF",
                                        'Ammonium sulphate at WWTP':"#f0f0fc",
                                        'compost, market for compost Metropolitan Area of Barcelona':"#5ec7fa"}, name=' ') +\
            #p9.facet_grid('Scenario ~ LCIA') +\
            p9.facet_wrap(["Scenario"]) +\
            p9.labels.ylab('Abiotic Resource Depletion (Tons Sb$_{eq}$ yr$^{-1}$)') +\
            p9.labels.xlab('') +\
            p9.theme(figure_size=(8, 6), axis_text_x=p9.element_text(rotation=90, hjust = 1))).draw(show=True)
fig.savefig('output/ADP_LCIA_yr_allscenarios.png', bbox_inches='tight')
